refactor: add rename lib file to config file
