+++ src/args.ts
@@ -1,140 +1,6 @@
 import { Command as Commander } from "commander"
+import type { BuildCommandOptions, DeprecatedGlobalOptions, GlobalOptions, Options } from "./config.js"
 import { getEnvVar } from "./util.js"
-
-/**
- * The options of cmake-ts that includes the command to run and the global options
- */
-export type Options = {
-  /**
-   * The command to run
-   */
-  command: Command
-} & GlobalOptions
-
-/**
- * The build command is a command that builds the project
- */
-export type BuildCommand = {
-  type: "build"
-  options: BuildCommandOptions
-}
-
-/**
- * The build config is a config that describes the build to run for cmake-ts
- */
-export type BuildCommandOptions = {
-  /**
-   * Named config(s) to build, which could be from default configs or the ones defined in the config file (package.json)
-   *
-   *  If no config is provided, it will build for the current runtime on the current system with the Release build type
-   *
-   * The default configs are combinations of `<Runtime>`, `<BuildType>`, `<Platform>`, and `<Architecture>`.
-   *
-   *  - `<Runtime>`: the runtime to use
-   *
-   *    e.g.: `node`, `electron`, `iojs`
-   *
-   *  - `<BuildType>`: the cmake build type (optimization level)
-   *
-   *    e.g.: `debug`, `release`, or `relwithdebinfo`
-   *
-   *  - `<Platform>`: the target platform
-   *
-   *    e.g.: `win32`, `linux`, `darwin`, `aix`, `android`, `freebsd`, `haiku`, `openbsd`, `sunos`, `cygwin`, `netbsd`
-   *
-   *  - `<Architecture>`: the target architecture
-   *
-   *    e.g.: `x64`, `arm64`, `ia32`, `arm`, `loong64`, `mips`, `mipsel`, `ppc`, `ppc64`, `riscv64`, `s390`, `s390x`
-   *
-   *   Any combination of `<BuildType>`, `<Runtime>`, `<Platform>`, and `<Architecture>` is valid. Some examples:
-   *
-   *    - `release`
-   *    - `debug`
-   *    - `relwithdebinfo`
-   *    - `node-release`
-   *    - `node-debug`
-   *    - `electron-release`
-   *    - `electron-debug`
-   *    - `win32-x64`
-   *    - `win32-x64-debug`
-   *    - `linux-x64-debug`
-   *    - `linux-x64-node-debug`
-   *    - `linux-x64-electron-release`
-   *    - `darwin-x64-node-release`
-   *    - `darwin-arm64-node-release`
-   *    - `darwin-arm64-electron-relwithdebinfo`
-   *
-   * You can also define your own configs in the config file (package.json).
-   *
-   *  - `<ConfigName>`: the name of the config
-   *
-   *    e.g.: `my-config`
-   *
-   *  The configs can also be in format of `named-<property>`, which builds the configs that match the property.
-   *
-   *  - `named-os`: build all the configs in the config file that have the same OS
-   *  - `named-os-dev`: build all the configs in the config file that have the same OS and `dev` is true
-   *  - `named-all`: build all the configs in the config file
-   *
-   *
-   *  The configs can be combined with `,` or multiple `--configs` flags. They will be merged together.
-   */
-  configs: string[]
-}
-
-/**
- * The help command is a command that shows the help message
- */
-export type HelpCommand = {
-  type: "help"
-}
-
-/**
- * A command is an object that describes the command to run for cmake-ts
- */
-export type Command = BuildCommand | HelpCommand | { type: "error" | "none" }
-
-/**
- * Global options are options that are available for all commands provided by the user as --option
- */
-export type GlobalOptions = {
-  /** Enable debug logging */
-  debug: boolean
-
-  /** Show help */
-  help: boolean
-}
-
-/**
- * Global options are options that are available for all commands provided by the user as --option
- * @deprecated Use the alternative options instead
- */
-export type DeprecatedGlobalOptions = {
-  /** Build all configurations
-   * @deprecated Use `build --config named-all` instead
-   */
-  all: boolean
-
-  /** Build only native configurations
-   * @deprecated Use `build` instead
-   */
-  nativeonly: boolean
-
-  /** Build only OS configurations
-   * @deprecated Use `build --config named-os` instead
-   */
-  osonly: boolean
-
-  /** Build only dev OS configurations
-   * @deprecated Use `build --config named-os-dev` instead
-   */
-  devOsOnly: boolean
-
-  /** Build only named configurations
-   * @deprecated Use `build --config <configs>` instead
-   */
-  namedConfigs?: string[]
-}
 
 /**
  * Parse the command line arguments and return the options.
