+++ src/main.ts
@@ -4,7 +4,7 @@ import { join, relative, resolve } from "path"
 import { copy, ensureDir, pathExists, readFile, readJson, remove, writeFile } from "fs-extra"
 import { parseArgs } from "./args.js"
 import { ArgumentBuilder } from "./argumentBuilder.js"
-import { type BuildConfigurations, parseBuildConfigs } from "./lib.js"
+import { type BuildConfigurations, parseBuildConfigs } from "./config.js"
 import { Logger } from "./logger.js"
 import { applyOverrides } from "./override.js"
 import { RuntimeDistribution } from "./runtimeDistribution.js"
