+++ src/runtimeDistribution.ts
@@ -2,8 +2,8 @@ import { extname, join as joinPath } from "path"
 import glob from "fast-glob"
 import { ensureDir, pathExists, readFile } from "fs-extra"
 import urlJoin from "url-join"
+import type { BuildConfiguration } from "./config.js"
 import { downloadFile, downloadTgz, downloadToString } from "./download.js"
-import type { BuildConfiguration } from "./lib.js"
 import { HOME_DIRECTORY, getPathsForConfig } from "./urlRegistry.js"
 import { stat } from "./util.js"
